Rename data_acquisition/gui_dialogs.py → dialog_request.py
Avoids confusion with app/gui_dialogs.py. The module contains only
the DialogRequest dataclass; the name now reflects that clearly.
Update imports in scan_manager.py and app/gui_dialogs.py.
